refactor: split stage and card components
